BE: Update price from Decimal128 to long

+++ server/src/main/java/com/FastFoodDelivery/dto/request/MenuItem/CreateMenuItemRequest.java
@@ -3,13 +3,11 @@ package com.FastFoodDelivery.dto.request.MenuItem;
 import lombok.Data;
 import org.bson.types.ObjectId;
 
-import java.util.Date;
-
 @Data
 public class CreateMenuItemRequest {
     private ObjectId restaurantId;
     private String name;
     private String description;
-    private double price;
+    private long price;
     private String imageUrl;
 }
